Move logging to own file
